+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.12.7] - ??? // FIXME lhellmann: put release date
+
+### Changed
+- Instead of using ping command to check connection, send ICMP packet directly
+
 ## [v0.12.16] - 2022-11-04
 
 ### Fixed